create project works

+++ public/js/dashboard.js
@@ -1,27 +1,6 @@
 // const newFormHandler = async (event) => {
 //   event.preventDefault();
 
-//   const name = document.querySelector('#project-name').value.trim();
-//   const description = document.querySelector('#project-desc').value.trim();
-//   const image = document.querySelector('#project-img').value.trim();
-
-//   if (name && needed_funding && description && image) {
-//     const response = await fetch(`/api/posts`, {
-//       method: 'POST',
-//       body: JSON.stringify({ name, description, image }),
-//       headers: {
-//         'Content-Type': 'application/json',
-//       },
-//     });
-
-//     if (response.ok) {
-//       document.location.replace('/dashboard');
-//     } else {
-//       alert('Failed to create post');
-//     }
-//   }
-// };
-
 // const delButtonHandler = async (event) => {
 //   if (event.target.hasAttribute('data-id')) {
 //     const id = event.target.getAttribute('data-id');
@@ -39,26 +18,23 @@
 // };
 
 // document
-//   .querySelector('.new-post-form')
-//   .addEventListener('submit', newFormHandler);
-
-// document
-//   .querySelector('.post-list')
+//   .querySelector('#delete-project')
 //   .addEventListener('click', delButtonHandler);
-
+// }
 
 let projectTitle;
 let projectDesc;
 let projectImg;
 let saveBtn;
-let deleteBtn;
+// let deleteBtn;
 
 if(window.location.pathname === '/dashboard') {
   projectTitle = document.querySelector('#project-title');
   projectDesc = document.querySelector('#project-desc');
   projectImg = document.querySelector('#project-img');
   saveBtn = document.querySelector('#save-project');
-  deleteBtn = document.querySelector('#delete-project');
+  // deleteBtn = document.querySelector('.delete-project');
+  // console.log(deleteBtn)
 }
 
 const saveProject = (event) => {
@@ -75,15 +51,38 @@ const saveProject = (event) => {
   })
 };
 
-const deleteProject = (event) => {
-  event.preventDefault();
-  fetch(`/api/projects/${id}`, {
-    method: 'DELETE',
-    headers: { 'Content-Type': 'application/json'},
-  })
+// const deleteProject = (event) => {
+//   event.preventDefault();
+//   console.log(event)
+//   fetch(`/api/projects/${id}`, {
+//     method: 'DELETE',
+//     headers: { 'Content-Type': 'application/json'},
+//   })
+// };
+
+const delButtonHandler = async (event) => {
+  if (event.target.hasAttribute('data-id')) {
+    const id = event.target.getAttribute('data-id');
+
+    const response = await fetch(`/api/projects/${id}`, {
+      method: 'DELETE',
+    });
+
+    if (response.ok) {
+      document.location.replace('/dashboard');
+    } else {
+      alert('Failed to delete project');
+    }
+  }
 };
 
-if(window.location.pathname === '/dashboard') {
+document
+  .querySelector('.delete-project')
+  .addEventListener('click', delButtonHandler);
+
+
+if(window.location.pathname === '/dashboard') 
+{
   saveBtn.addEventListener('click', saveProject);
-  deleteBtn.addEventListener('click', deleteProject);
+  // deleteBtn.addEventListener('click', deleteProject);
 }
